Update script.js
The OpenWeatherMap API key I provided in the previous example is no longer valid. To use the OpenWeatherMap API, you will need to sign up for a free API key on their website. Once you have an API key, you can replace YOUR_API_KEY with your actual API key in the JavaScript code to make the API calls to their servers.

Additionally, it's important to note that the OpenWeatherMap API has usage limits, so you will need to be mindful of the number of requests you are making and ensure that you are not exceeding the usage limits. If you exceed the usage limits, you may receive an error message or your API key may be blocked.

Also, the API endpoint may change due to their policy change, thus you need to check their documentation for the endpoints to use.